requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
dill==0.3.4
fonttools==4.28.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
importlib-metadata==4.10.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.0
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31.post2005241907
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
